% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-ratios.R
\name{AverageRecurringRevenue}
\alias{AverageRecurringRevenue}
\title{\code{AverageRecurringRevenue}}
\usage{
AverageRecurringRevenue(data, ratio = TRUE)
}
\arguments{
\item{data}{A \code{MetricData} object.}

\item{ratio}{If \code{TRUE}, the statistic is returned as a ratio.}
}
\value{
A named vector if \code{cohort.type} is set to \code{"None"} or \code{"Preceding"}, or,
  a \code{matrix}. This will contain a number of attributes includeing:
\enumerate{
  \item \code{denominator} cohort.typed when \code{ratio} is \code{TRUE}.
  \item \code{numerator} cohort.typed when \code{ratio} is \code{TRUE}.
  \item The \code{detail} cohort.typed when \code{ratio} is \code{TRUE}.
  }
}
\description{
The average recurring revenue provided by each customer.
}
\details{
Calculated based on all the customer immediately prior to the end of the
previous period.
}
