% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiescharting.R
\name{cohortHeatmap}
\alias{cohortHeatmap}
\title{\code{cohortHeatmap}}
\usage{
cohortHeatmap(x, series.hover, ...)
}
\arguments{
\item{x}{A \code{matrix}}

\item{series.hover}{The description of the value that appears in each cell;
used in the hover tooltips}

\item{...}{Additional arguments to be passed to lower level functions.}
}
\value{
A plotly heatmap
}
\description{
Internal heatmap function.
}
