% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{convertToMatrix}
\alias{convertToMatrix}
\title{#' @importFrom flipTime AsDate Period Periods
#' @importFrom lubridate floor_date
diagRectangular <- function(x, data)
{
    by <- attr(data, "by")
    subscription.length <- attr(data, "subscription.length")
    unit <- Periods(1, subscription.length)
    row.dts <- AsDate(rownames(x))
    rows <- floor_date(row.dts, subscription.length) + unit
    rowm <- matrix(rows, nrow(x), ncol(x))                   
    cols <- AsDate(colnames(x))
    colm <- matrix(cols, nrow(x), ncol(x), byrow = TRUE)    
    m <- colm == rowm
    out <- x[m]
    names(out) <- Period(row.dts [rowSums(m) != 0] + unit, by)
    out
}}
\usage{
convertToMatrix(results, statistic)
}
\description{
#' @importFrom flipTime AsDate Period Periods
#' @importFrom lubridate floor_date
diagRectangular <- function(x, data)
{
    by <- attr(data, "by")
    subscription.length <- attr(data, "subscription.length")
    unit <- Periods(1, subscription.length)
    row.dts <- AsDate(rownames(x))
    rows <- floor_date(row.dts, subscription.length) + unit
    rowm <- matrix(rows, nrow(x), ncol(x))                   
    cols <- AsDate(colnames(x))
    colm <- matrix(cols, nrow(x), ncol(x), byrow = TRUE)    
    m <- colm == rowm
    out <- x[m]
    names(out) <- Period(row.dts [rowSums(m) != 0] + unit, by)
    out
}
}
