% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenuemetric.R
\name{createOutput}
\alias{createOutput}
\title{Create a standardized object}
\usage{
createOutput(x, class.name, calculation, name)
}
\arguments{
\item{x}{The statistic to be displayed.}

\item{class.name}{Used to determine how to plot the data}

\item{calculation}{The detail}

\item{name}{Used in plotting}
}
\description{
Create a standardized object
}
