% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-ratios.R
\name{CustomerChurn}
\alias{CustomerChurn}
\title{\code{CustomerChurn}}
\usage{
CustomerChurn(data, use = "Aggregate", ratio = TRUE)
}
\arguments{
\item{data}{A \code{MetricData} object.}

\item{use}{The data to be used in the calculation:
\enumerate{
  \item \code{"Aggregate"} All data is used in the calculation for each period.
  \item \code{"Initial"} Data for the initial subscription period is used. 
  \item \code{"Cohort"} The statistic is computed for each cohort by each 
  subscription period.
}}

\item{ratio}{If \code{TRUE}, the statistic is returned as a ratio.}
}
\value{
A named vector if \code{use} is set to \code{"Aggregate"} or \code{"Initial"}, or,
  a \code{matrix}. This will contain a number of attributes includeing:
\enumerate{
  \item \code{denominator} Used when \code{ratio} is \code{TRUE}.
  \item \code{numerator} Used when \code{ratio} is \code{TRUE}.
  \item The \code{detail} used when \code{ratio} is \code{TRUE}.
  }
}
\description{
The percentage or number of customers who could churn in a period that did churn.
}
\details{
The definition of churn is is not the most widely used definition.
A more common definition is the proportion of customers that churned. 
These two measures differ in some situations:
\enumerate{
  \item Incomplete periods. The traditional metric can only be used with completed metrics.
This function computes churn for periods that are not yet complete, by using the dat
for the customers that were have been up for renewal prior to the point at which the
analysis is conducted.
  \item Contracts that extend over multiple subsscription lengths. E.g., a two year
contract for a product that typically has a one year subscription. This function
ignores such contracts in their first year (as the customer cannot churn). Thus, 
this function will show a higher churn rate than the more traditional churn.
  \item Contracts that start and end with a subscription-length period. For example, 
if subscriptions are yearly, and a customer starts on the 1st of Jan and churns on the 
30th of December, they will not appear in the churn statistics in the traditional
metric. They do appear in the calculations in this function.
}
}
