% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-ratios.R
\name{CustomerRetention}
\alias{CustomerRetention}
\title{\code{CustomerRetention}}
\usage{
CustomerRetention(data, ratio = TRUE)
}
\arguments{
\item{data}{A \code{MetricData} object.}

\item{ratio}{If \code{TRUE}, the statistic is returned as a ratio.}
}
\value{
A named vector if \code{cohort.type} is set to \code{"None"} or \code{"Preceding"}, or,
  a \code{matrix}. This will contain a number of attributes includeing:
\enumerate{
  \item \code{denominator} cohort.typed when \code{ratio} is \code{TRUE}.
  \item \code{numerator} cohort.typed when \code{ratio} is \code{TRUE}.
  \item The \code{detail} cohort.typed when \code{ratio} is \code{TRUE}.
  }
}
\description{
The number of proportion of customers that are retained.
}
\details{
Based on those whose contracts were
up for renewal.
}
