% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenuemetric.R
\name{Denominator}
\alias{Denominator}
\title{\code{Denominator}}
\usage{
Denominator(x)
}
\arguments{
\item{x}{A \code{RevenueMetric} object.}
}
\description{
Extracts the denominator from a RevenueMetric object (or a component of in some situations).
}
