% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthaccounting.R
\name{GrowthAccounting}
\alias{GrowthAccounting}
\title{\code{GrowthAccounting}}
\usage{
GrowthAccounting(data, by = "year", ...)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{by}{The time period used in defining the cohorts: "year", "quarter", "month", "week", day".}

\item{...}{Additional arguments to be passed to lower level functions.}
}
\value{
A matrix
}
\description{
Computes the growth within accounts
}
