% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialchurn.R
\name{InitialCustomerChurn}
\alias{InitialCustomerChurn}
\title{\code{InitialCustomerChurn}}
\usage{
InitialCustomerChurn(data)
}
\arguments{
\item{data}{A \code{MetricData} object.}
}
\value{
A named vector showing churn.
}
\description{
The percentage of customers in their first period of a subscription who don't renew for a 
second subscription. See \code{CustomerChurn}.
}
