% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{mergersWithDates}
\alias{mergersWithDates}
\title{Adds a column of the tim period at which the mergers occurred}
\usage{
mergersWithDates(data)
}
\description{
Adds a column of the tim period at which the mergers occurred
}
