% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricdata.R
\name{MetricData}
\alias{MetricData}
\title{\code{MetricData}}
\usage{
MetricData(
  value,
  from,
  to,
  start,
  end,
  id,
  subscription.length,
  by,
  mergers,
  trim.id
)
}
\arguments{
\item{value}{A vector of containing the revenue per transaction.}

\item{from}{A vector of class \code{POSIXct} or \code{POSIXt},
recording the date and time each subscription commences.}

\item{to}{A vector of class \code{POSIXct} or \code{POSIXt},
recording the date and time each subscription ends}

\item{start}{The date at which the analysis outputs should
commence. By default, the earliest date recorded in
\code{from}.}

\item{end}{The date at which the analysis ends, which is used to
determine churn.  By default, the most recent date recorded in
\code{from}.}

\item{id}{A vector of \code{character}, unique identifier for
subscribers that made the transactions (e.g., email addresses,
names, subscriber keys).}

\item{subscription.length}{The time unit that describes the
subscription length: \code{year} to view the data by year,
\code{quarter}, and \code{month}. This is assumed to be the
billing period when determining if subscribers have churned or
not.}

\item{by}{The time unit that describes the
subscription length: \code{year} to view the data by year,
\code{quarter}, and \code{month}. This is assumed to be the
billing period when determining if subscribers have churned or
not.}

\item{mergers}{A data frame with two variables 'id' and 'id.to'. 
'id' contains the ids of companies that appeara, based on their data,
to have churned, but have in fact merged with the corresponding 'id.to'.}

\item{trim.id}{The maximum length of the strings to be used showing
ID names (used to avoid situations where string names are so
long as to make reading of tables impossible.}
}
\value{
A \code{\link{data.frame}} where the rows represent invoice lines
    and the variables are: 
    \itemize{
       \code{id} {The unique identifier.}  
       \code{value} {The total fee/price for a subscription.}
       \code{from} {The commencement date of a subscription.}
       \code{recurring.value} {The value divided by proportion of the typicaly invoice period
       that was covered by the invoice. There are some rounding error issues (e.g., leap years,
       inconsistencies in how people enter data)}.
   }
   Additional information is included in attributes
}
\description{
Cleans and tidies data for use in revenue metric calculations.
}
