% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{NetRecurringRevenueChurn}
\alias{NetRecurringRevenueChurn}
\title{\code{NetRecurringRevenueChurn}}
\usage{
NetRecurringRevenueChurn(data, use = "aggregate")
}
\arguments{
\item{data}{A \code{MetricData} object.}

\item{use}{The data to be used in the calculation: 'aggregate, "initial', or 'chort'.}
}
\value{
A named vector if \code{use = 'aggregate'} or \code{use = 'initial'}, or,
  a \code{matrix}. This will contain a number of attributes includeing:
\enumerate{
  \item The \code{denominator}. Where the statistic is a ratio, this is it's
  denominator. Otherwise, it is typically total recurring revenue.
  \item The \code{numerator}. Where the statistic is a ratio, this is it's
  numerator Otherwise, it is typically the number of customers.
  }
}
\description{
The change in the recurring revenue over time. A positive value inducates
that [Expansion()] + [Contraction()] > [RecurringRevenueChurn()]
}
\details{
Calculated based on all the customer immediately prior to the end of the
previous period.
}
