% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-ratios.R
\name{NetRecurringRevenueRetention}
\alias{NetRecurringRevenueRetention}
\title{\code{NetRecurringRevenueRetention}}
\usage{
NetRecurringRevenueRetention(data)
}
\arguments{
\item{data}{A \code{MetricData} object.}
}
\value{
A named vector if \code{cohort.type} is set to \code{"None"} or \code{"Preceding"}, or,
  a \code{matrix}. This will contain a number of attributes includeing:
\enumerate{
  \item \code{denominator} cohort.typed when \code{ratio} is \code{TRUE}.
  \item \code{numerator} cohort.typed when \code{ratio} is \code{TRUE}.
  \item The \code{detail} cohort.typed when \code{ratio} is \code{TRUE}.
  }
}
\description{
The change in the recurring revenue over time. A positive value indicates
that the company is growing even without adding any new sales.
that [Expansion()] + [Contraction()] > [RecurringRevenueChurn()].
}
\details{
Calculated based on all the customer immediately prior to the end of the
previous period. Note that this is not the commplement of [RecurringRevenueChurn()],
as that metric only takes into account churn, whereas this metric also 
includes [Expansion()] and [Contraction()]
}
