% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-ratios.R
\name{NewRecurringRevenueGrowth}
\alias{NewRecurringRevenueGrowth}
\title{\code{NewRecurringRevenueGrowth}}
\usage{
NewRecurringRevenueGrowth(data)
}
\arguments{
\item{data}{A \code{MetricData} object.}
}
\value{
A named vector if \code{cohort.type} is set to \code{"None"} or \code{"Preceding"}, or,
  a \code{matrix}. This will contain a number of attributes includeing:
\enumerate{
  \item \code{denominator} cohort.typed when \code{ratio} is \code{TRUE}.
  \item \code{numerator} cohort.typed when \code{ratio} is \code{TRUE}.
  \item The \code{detail} cohort.typed when \code{ratio} is \code{TRUE}.
  }
}
\description{
The growth (as a percentage) in sales.
}
\details{
Calculated as any sale in a period made
to a customer who was not a customer at the beginning
of the period. This includes resurrections.
Where the period has been set to something other than
"year" the results are annualized (e.g., monthly data
is multiplier by 12).

Note that when the period is changed, this can
affect the average (e.g., the trend line), as
for example, yearly data will be weighted towards
data with bigger baseline sales.  XXX
}
