% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenuemetric.R
\name{Numerator}
\alias{Numerator}
\title{\code{Numerator}}
\usage{
Numerator(x)
}
\arguments{
\item{x}{A \code{RevenueMetric} object.}
}
\description{
Extracts the numerator from a RevenueMetric object (or a component of in some situations).
}
