% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrowthAccountingOld.R
\name{plot.GrowthAccounting}
\alias{plot.GrowthAccounting}
\title{#' @export
print.GrowthAccountingOld <- function(x, ...)
{
    cat('Growth Accounting calculations\n\n')
    cat(paste0('\nNumber of entities: ', x$n.id, '\n'))
    cat('\nRevenue\n')
    print(x$Revenue)
    cat('\nRevenue Growth\n')
    pre <- x$Revenue[-x$n.periods]
    revenue.growth <- round((x$Revenue[-1] - pre) / pre * 100, 0)
    names(revenue.growth) <- x$periods[-1]
    print(revenue.growth)
    cat('\nGrowth\n')
    print(x$Growth)
    cat('\nTable\n')
    print(x$Table)
    #cat('\n')
}}
\usage{
\method{plot}{GrowthAccounting}(x, ...)
}
\description{
#' @export
print.GrowthAccountingOld <- function(x, ...)
{
    cat('Growth Accounting calculations\n\n')
    cat(paste0('\nNumber of entities: ', x$n.id, '\n'))
    cat('\nRevenue\n')
    print(x$Revenue)
    cat('\nRevenue Growth\n')
    pre <- x$Revenue[-x$n.periods]
    revenue.growth <- round((x$Revenue[-1] - pre) / pre * 100, 0)
    names(revenue.growth) <- x$periods[-1]
    print(revenue.growth)
    cat('\nGrowth\n')
    print(x$Growth)
    cat('\nTable\n')
    print(x$Table)
    #cat('\n')
}
}
