% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricdata.R
\name{processMergers}
\alias{processMergers}
\title{Adds a column of the tim period at which the mergers occurred}
\usage{
processMergers(id, to, by, mergers)
}
\description{
Adds a column of the tim period at which the mergers occurred
}
