% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{q.invoice.lines}
\alias{q.invoice.lines}
\title{Variables from a companies table of invoice lines.}
\format{
data.frame

data.frame
}
\description{
Variables from a companies table of invoice lines.

Information about Q's subscriber history.
}
\keyword{data}
