% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{q.invoice.lines.short}
\alias{q.invoice.lines.short}
\title{Variables from a companies table of invoice lines.}
\format{
data.frame
}
\description{
Variables from a companies table of invoice lines.
A sub-sample of q.invoice.lines
}
\keyword{data}
