% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-univariate.R
\name{RecurringRevenue}
\alias{RecurringRevenue}
\title{\code{RecurringRevenue}}
\usage{
RecurringRevenue(data, use = "Aggregate")
}
\arguments{
\item{data}{A \code{MetricData} object.}

\item{use}{The data to be used in the calculation:
\enumerate{
  \item \code{"Aggregate"} All data is used in the calculation for each period.
  \item \code{"Initial"} Data for the initial subscription period is used. 
  \item \code{"Cohort"} The statistic is computed for each cohort by each 
  subscription period.
}}
}
\value{
A vector showing the recurring revenue by time points.
}
\description{
Computes recurring revenue at the end of the period (i.e., as of
the instant before the next time periods starts). Where the time period isn't complete, 
the recurring revenue is calculated up until the instant before attr(data, "end").
}
\details{
For example, if with yearly data, if a customer has an annual license for $1,000
from 1 February to 1 February of the next year, and a second license which starts on 1 January of the last year
with a prorated price of $1000/12, then the recurring revenue is $2,000.
}
