% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurringrevenuebycohort.R
\name{RecurringRevenueByCohort}
\alias{RecurringRevenueByCohort}
\title{\code{RecurringRevenueByCohort}}
\usage{
RecurringRevenueByCohort(
  data,
  cohort.by = "year",
  by = "year",
  remove.last = TRUE,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{cohort.by}{The time period used in defining the cohorts: "
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, 
and \code{"day"}.}

\item{by}{The time period to aggregate the dates by: 
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, 
and \code{"day"}.}

\item{remove.last}{Remove the final period (this is useful if it is likely incomplete.).}

\item{...}{Additional arguments to be passed to lower level functions.}
}
\value{
A table.
}
\description{
Revenue by year and start year, as a stacked revenue chart.
}
