% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churnbycohort.R
\name{RecurringRevenueChurnByCohort}
\alias{RecurringRevenueChurnByCohort}
\title{\code{RecurringRevenueChurnByCohort}}
\usage{
RecurringRevenueChurnByCohort(data, by, remove.last = FALSE, ...)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{by}{The time period to aggregate the dates by: 
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, 
and \code{"day"}.}

\item{remove.last}{Remove the final period (this is useful if it is likely incomplete.).}

\item{...}{Additional arguments to be passed to lower level functions.}
}
\value{
A \code{\link{matrix}}
}
\description{
Computes the recurring revenue churn by time period and time-based cohort, where the time
periods are determined by the \code{attr(data, "subscription.length")}.
}
