% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retentionbycohort.R
\name{RecurringRevenueRetentionByCohort}
\alias{RecurringRevenueRetentionByCohort}
\title{\code{RecurringRevenueChurn}}
\usage{
RecurringRevenueRetentionByCohort(data)
}
\arguments{
\item{data}{A \code{RevenuMetric} object.}
}
\value{
A named vector showing churn, a plot, or an object showing detail.
}
\description{
Lost revenue duee to churned customers as a percentage of total recurring revenue.
}
