% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenuemetric.R
\name{RevenueMetric}
\alias{RevenueMetric}
\title{\code{RevenueMetric}}
\usage{
RevenueMetric(
  FUN = "Acquisition",
  output = c("Plot", "Table", "Detail")[1],
  by = c("day", "month", "quarter", "year")[4],
  cohort.type = "None",
  cohort.period = "year",
  id,
  value,
  from,
  to,
  start = as.Date(min(from)),
  end = as.Date(max(from)),
  subscription.length = "year",
  subset = rep(TRUE, length(id)),
  profiling = NULL,
  trim.id = 50,
  mergers = NULL,
  ...
)
}
\arguments{
\item{FUN}{A function that calculates a metric}

\item{output}{Whether to output as a Plot, Table, or Detail}

\item{by}{The unit of time to report on ("day", "month", "quarter", "year").}

\item{cohort.type}{How cohorts are to be used when performing the analysis
\code{by} setting.
\itemize{
 \code{"None"} {All customers are cohort.typed in calculations.}
 \code{"New"} {Customers added in the preceding \code{by}.}
 \code{"Calendar"} {Calender time periods are used. For example, if \code{cohort.period} is set
 to "year" then calendar years are used as cohorts.}
 \code{"Tenure"} {Tenure based time periods are used. For example, if \code{cohort.period} is set
 to "year", then cohorts are based on the number of years since a customeer first signed up.
 then calendar years are used as cohorts.}
}}

\item{cohort.period}{The period of aggregation to be used, with options of \code{"week"}, \code{"month"},
\code{"quarter"}, and \code{"year"}. This parameter is only used when \code{cohort.type} is \code{"Calendar"}
or \code{"Tenure"}}

\item{id}{A vector of \code{character}, unique identifier for
subscribers that made the transactions (e.g., email addresses,
names, subscriber keys).}

\item{value}{A vector of containing the revenue per transaction.}

\item{from}{A vector of class \code{POSIXct} or \code{POSIXt},
recording the date and time each subscription commences.}

\item{to}{A vector of class \code{POSIXct} or \code{POSIXt},
recording the date and time each subscription ends}

\item{start}{The date at which the analysis outputs should
commence. By default, the earliest date recorded in
\code{from}.}

\item{end}{The date at which the analysis ends, which is used to
determine churn.  By default, the most recent date recorded in
\code{from}.}

\item{subscription.length}{The time unit that describes the
subscription length: \code{year} to view the data by year,
\code{quarter}, and \code{month}. This is assumed to be the
billing period when determining if subscribers have churned or
not.}

\item{subset}{An optional vector specifying a subset of
observations to be used in the calculations}

\item{profiling}{Separate analyses are conducted among each unique combination of these variables.}

\item{trim.id}{The maximum length of the strings to be used showing
    ID names (used to avoid situations where string names are so
    long as to make reading of tables impossible.
\code{"quarter"}, and \code{"year"}.}

\item{mergers}{A data frame with two variables 'id' and 'id.to'. 
'id' contains the ids of companies that appeara, based on their data,
to have churned, but have in fact merged with the corresponding 'id.to'.}

\item{...}{Additional arguments to be passed to lower level functions.}
}
\value{
A plotly plot#?
}
\description{
Creates a small multiple plot by sub-groups
}
\details{
The \code{cohort.type} parameter does not have the transitive properties
that many assume. For example, it's not always the case that \code{"New"} is 
equivalent to a diagonal from \code{"Calendar"}, or that \code{"None"} is 
the column-sums of \code{"Calendar"}. 

Consider as an example \code{CustomerChurn} where
all analysis is being done with the unit of \code{"year"}. A customer that 
purchases a contract less than the length of a year will appear below the
(possibly offset) diagonal in \code{"Calendar"} and if they renew will be in 
the denominator for the following year of \code{"New"}, but not if they do 
not renew.

Similarly, with \code{RecurringRevenueChurn}, \code{"New"} in an incomplete final
 period will compare to the same period in the previous year, whereas the bottom
 row of \code{"Calendar"}  will use the entire calendar yearear as the base, and
 thus look much better (in a misleading way).
}
