% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{\code{calculate}}
\usage{
calculate(data, components, volume, use)
}
\arguments{
\item{data}{A \code{MetricData} object.}

\item{components}{A character indicating the revenue metric component to use; one
of 'churn', 'expansion', 'contraction', 'retention', and 'net retention'.}

\item{volume}{If TRUE, recurring revenue rather than customers is used in calculations.}

\item{use}{The data to be used in the calculation: 'aggregate, "initial', or 'chort'.}
}
\value{
A named vector showing churn.
}
\description{
The percentage of customers who could churn in a period that did churn.
Note that this is not the most widely used definition. A more common definition
is the proportion of customers that churned. These two measures differ in some  situations:
1. Multi-year contracts.
2. Incomplete periods (e.g., the conventional metric can't be used in a quarter until it
has been completed.
3. Where a customer joins and churns withhin a single period (ignored in the traditional definition)
}
