% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{\code{calculate}}
\usage{
calculate(data, components, volume, use)
}
\arguments{
\item{data}{A \code{MetricData} object.}

\item{components}{A character indicating the revenue metric component to use; one
of 'churn', 'expansion', 'contraction', 'retention', and 'net retention'.}

\item{volume}{If TRUE, recurring revenue rather than customers is used in calculations.}

\item{use}{The data to be used in the calculation: 'aggregate, "initial', or 'chort'.}
}
\value{
A named vector showing churn.
}
\description{
Calculates main metrics
}
