% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{\code{calculate}}
\usage{
calculate(data, initial.only = FALSE, by.period = FALSE, volume = TRUE)
}
\arguments{
\item{data}{A \code{MetricData} object.}

\item{initial.only}{If TRUE, only calculates retention for the first opportunity for
renewal}

\item{by.period}{If TRUE, calculates one figure for each cohort (rather than by period)}
}
\value{
A named vector showing churn.
}
\description{
The percentage of customers who could churn in a period that did churn.
Note that this is not the most widely used definition. A more common definition
is the proportion of customers that churned. These two measures differ in some  situations:
1. Multi-year contracts.
2. Incomplete periods (e.g., the conventional metric can't be used in a quarter until it
has been completed.
3. Where a customer joins and churns withhin a single period (ignored in the traditional definition)
}
