% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{CustomerChurn}
\alias{CustomerChurn}
\title{\code{CustomerChurn}}
\usage{
CustomerChurn(data, use = "aggregate")
}
\arguments{
\item{data}{A \code{MetricData} object.}

\item{use}{The data to be used in the calculation: 'aggregate, "initial', or 'chort'.}
}
\value{
A named vector if \code{use = 'aggregate'} or \code{use = 'initial'}, or,
  a \code{matrix}. This will contain a number of attributes includeing:
\enumerate{
  \item The \code{denominator}. Where the statistic is a ratio, this is it's
  denominator. Otherwise, it is typically total recurring revenue.
  \item The \code{numerator}. Where the statistic is a ratio, this is it's
  numerator Otherwise, it is typically the number of customers.
  }
}
\description{
The percentage of customers who could churn in a period that did churn.
}
\details{
The definition of churn is is not the most widely used definition.
A more common definition is the proportion of customers that churned. 
These two measures differ in some situations:
\enumerate{
  \item Incomplete periods. The traditional metric can only be used with completed metrics.
This function computes churn for periods that are not yet complete, by using the dat
for the customers that were have been up for renewal prior to the point at which the
analysis is conducted.
  \item Contracts that extend over multiple subsscription lengths. E.g., a two year
contract for a product that typically has a one year subscription. This function
ignores such contracts in their first year (as the customer cannot churn). Thus, 
this function will show a higher churn rate than the more traditional churn.
  \item Contracts that start and end with a subscription-length period. For example, 
if subscriptions are yearly, and a customer starts on the 1st of Jan and churns on the 
30th of December, they will not appear in the churn statistics in the traditional
metric. They do appear in the calculations in this function.
}
}
