% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{Customers}
\alias{Customers}
\title{\code{Customers}}
\usage{
Customers(data)
}
\arguments{
\item{data}{A ...}

\item{...}{Additional arguments to be passed to lower level functions.}
}
\value{
A named vector showing the number of customers.
}
\description{
The number of customers (i.e., unique values of data$id, with positive value at a given pont in time).
}
