% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-univariate.R
\name{Customers}
\alias{Customers}
\title{\code{Customers}}
\usage{
Customers(data)
}
\arguments{
\item{data}{A \code{MetricData} object.}
}
\value{
A vector showing the recurring revenue by time points.
}
\description{
Number of customers by time period..
}
\details{
For example, if with yearly data, if a customer has an annual license for $1,000
from 1 February to 1 February of the next year, and a second license which starts on 1 January of the last year
with a prorated price of $1000/12, then the recurring revenue is $2,000.
}
