% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthaccounting.R
\name{GrowthAccounting}
\alias{GrowthAccounting}
\title{\code{GrowthAccounting}}
\usage{
GrowthAccounting(data, small = 0.1)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{small}{A proportion. Expansion less than or equal to this  is classified as minor expansion (e.g., price rises) or
an increase in license sales in a larger company.}
}
\value{
A matrix
}
\description{
Computes the growth within accounts
}
