% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialchurn.R
\name{InitialCustomerChurn}
\alias{InitialCustomerChurn}
\title{#' @importFrom flipTime AsDate Periods 
selectChurnData <- function(x, subscription.length, by)
{
    n.subscribers <- attr(x, "n.subscribers")
    row.dates <- AsDate(rownames(x))
    col.dates <- AsDate(colnames(x))
    initial.possible.renewal <- row.dates + Periods(1, subscription.length) 
    m <- match(initial.possible.renewal, col.dates)
    k <- length(row.dates)
    churn <- rep(NA, k)
    base <- rep(0, k)
    names(base) <- names(churn) <- Period(initial.possible.renewal, by)
    not.missing <-!is.na(m) 
    lookups <- cbind(1:k, m)[not.missing, , drop = FALSE]
    churn[not.missing] <- x[lookups]
    base[not.missing] <- n.subscribers[lookups]
    attr(churn, "n.subscribers") <- base
    churn
}
InitialCustomerChurn}
\usage{
InitialCustomerChurn(data, by, remove.last, ...)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}

\item{by}{The time period to aggregate the dates by: 
\code{"year"}, \code{"quarter"}, \code{"month"}, \code{"week"}, 
and \code{"day"}.}

\item{remove.last}{Remove the final period (as usually is incomplete).}

\item{...}{Additional arguments to be passed to lower level functions.}
}
\description{
Computes the churn rate for people in their first completed
renewal period.
}
\details{
For example, if the subscriptions are annual, computes 
proportion of people in 2012 that churned in 2013. Customers that had
an incomplete year that falls within a calendar year will be excluded
from this metric (e.g., a license from 1 January to 1 December 2012  will
not appear in this calculation, but their renewal year may instead). 
This calculation is taken from the one-off diagonal of 
\code{CustomerChurnByCohort}.
}
