% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-ratios.R
\name{NetRecurringRevenueRetention}
\alias{NetRecurringRevenueRetention}
\title{\code{NetRecurringRevenueRetention}}
\usage{
NetRecurringRevenueRetention(data, use = "Aggregate", ratio = TRUE)
}
\arguments{
\item{data}{A \code{MetricData} object.}

\item{use}{The data to be used in the calculation:
\enumerate{
  \item \code{"Aggregate"} All data is used in the calculation for each period.
  \item \code{"Initial"} Data for the initial subscription period is used. 
  \item \code{"Cohort"} The statistic is computed for each cohort by each 
  subscription period.
}}

\item{ratio}{If \code{TRUE}, the statistic is returned as a ratio.}
}
\value{
A named vector if \code{use} is set to \code{"Aggregate"} or \code{"Initial"}, or,
  a \code{matrix}. This will contain a number of attributes includeing:
\enumerate{
  \item \code{denominator} Used when \code{ratio} is \code{TRUE}.
  \item \code{numerator} Used when \code{ratio} is \code{TRUE}.
  \item The \code{detail} used when \code{ratio} is \code{TRUE}.
  }
}
\description{
The change in the recurring revenue over time. A positive value inducates
that [Expansion()] + [Contraction()] > [RecurringRevenueChurn()].
}
\details{
Calculated based on all the customer immediately prior to the end of the
previous period. Note that this is not the commplement of [RecurringRevenueChurn()],
as that metric only takes into account churn, whereas this metric also 
includes [Expansion()] and [Contraction()]
}
