% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthaccounting.R
\name{plot.GrowthAccounting}
\alias{plot.GrowthAccounting}
\title{plot.GrowthAccounting}
\usage{
\method{plot}{GrowthAccounting}(x, ...)
}
\arguments{
\item{x}{The metric to be plotted.}

\item{...}{Othe parameters.}
}
\description{
plot.GrowthAccounting
}
