% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurringrevenue.R
\name{RecurringRevenue}
\alias{RecurringRevenue}
\title{\code{RecurringRevenue}}
\usage{
RecurringRevenue(data)
}
\arguments{
\item{data}{A \code{data.frame} that has the same variables as a \code{RevenueData} object.}
}
\value{
A vector showing the recurring revenue by time points.
}
\description{
Computes recurring revenue at the end of the period (i.e., as of
the instant before the next time periods starts). Where the time period isn't ocmplete, 
the recurring revenue is calculated up until the instant before attr(data, "end").
}
\details{
For exmaple, if with yearly data, if a customer has an annual license for $1,000
from 1 February to 1 February of the next year, and a second license which starts on 1 January of the last year
with a prorated price of $1000/12, then the recurring revenue is $2,000.
}
