% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{RecurringRevenueChurn}
\alias{RecurringRevenueChurn}
\title{\code{RecurringRevenueChurn}}
\usage{
RecurringRevenueChurn(data, use)
}
\arguments{
\item{data}{A \code{MetricData} object.}

\item{use}{The data to be used in the calculation: 'aggregate, "initial', or 'chort'.}
}
\value{
A named vector if \code{use = 'aggregate'} or \code{use = 'initial'}, or,
  a \code{matrix}. This will contain a number of attributes includeing:
\enumerate{
  \item The \code{denominator}. Where the statistic is a ratio, this is it's
  denominator. Otherwise, it is typically total recurring revenue.
  \item The \code{numerator}. Where the statistic is a ratio, this is it's
  numerator Otherwise, it is typically the number of customers.
  }
}
\description{
Lost revenue due to churned customers as a percentage of total recurring revenue.
}
\details{
The calculation is based on all customers priorimmedidately prior to the 
end of the period. Note that this definition is not merely the recurrning revenue-weighted
equivalent of [CustomerChurn()], as [CustomerChurn()] is based on customers that could
have churned at any stage in the period.
}
